Update Dockerfile to use debian 'bookworm' and modify installation sequences
The Dockerfile has been updated to use the 'bookworm' version of debian from 'buster', along with the 'bookworm' version of golang. The installation sequences are also modified for better readability and efficiency. Apt update and install commands are grouped to make use of docker layer caching and remove redundant layers